==> 2_model_cal_val.Rmd <==
---
title: "__FFS and Spatial CV"
author: "Simon Topp"
date: "8/27/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---
# This script takes the munged optical and landscape data and runs it through various feature selection algorithms to determine the inputs for the final models.

```{r setup, eval = F}
library(tidyverse)
library(feather)
library(viridis)
library(knitr)
library(sf)
library(rgdal)
library(maps)
library(magrittr)
library(mlbench)
library(caret)
library(doParallel)
library(furrr)
library(lubridate)
library(groupdata2)
library(CAST)
library(mapview)
library(onehot)
library(Metrics)
library(kableExtra)
library(ggpmisc)
```

#Attempt at FFS Using xgboost and spatial cross validation.

```{r}
## Do some final munging/feature engineering
# srMunged <- read_feather('out/srMunged.feather') %>%
#   filter(parameter == 'secchi',
#          value <= 10,
#          !is.na(CatAreaSqKm)) %>% #100 obs that didn't match up with LakeCat
#   mutate(COMID = as.character(COMID),
#          pctForest2006 = PctDecid2006Cat + PctConif2006Cat + PctMxFst2006Cat,
#          pctUrban2006 = PctUrbMd2006Cat + PctUrbHi2006Cat,
#          pctWetland2006 = PctWdWet2006Cat + PctHbWet2006Cat,
#          areasqkm = round(areasqkm, 1),
#          meandused = round(meandused, 1),
#          inStreamCat = as.numeric(inStreamCat),
#          #Correct the reflectance values based on distribution comparisons (see line 358 in 1_nhd_join_and_munge.Rmd)
#          blue = ifelse(sat == '5', blue - .84, ifelse(sat == '8', blue + 76.3, blue)),
#          green = ifelse(sat == '5', green - 29.2, ifelse(sat == '8', green + 46.9, green)),
#          nir = ifelse(sat == '5', nir - 34.3, ifelse(sat == '8', nir + 127, nir)),
#          red = ifelse(sat == '5', red - 26.1, ifelse(sat == '8', red + 51.5, red)),
#          NR = nir/red,
#          BG = blue/green,
#          dWL = fui.hue(red, green, blue)) %>%
#   filter_at(vars(nir, red, blue, green), all_vars(.>0 & .<2000)) %>%
#   rowwise() %>%
#   mutate(saturation = rgb2hsv(r=red, g=green, b=blue, maxColorValue = 2000)[2],
#          brightness = rgb2hsv(r=red,g= green, b= blue, maxColorValue = 2000)[3]) %>%
#   ungroup()


srMunged <- read_feather('out/srMunged.feather') %>%
  filter(parameter == 'secchi',
         value <= 10,
         !is.na(CatAreaSqKm)) %>% #100 obs that didn't match up with LakeCat
  mutate(COMID = as.character(COMID),
         pctForest2006 = PctDecid2006Cat + PctConif2006Cat + PctMxFst2006Cat,
         pctUrban2006 = PctUrbMd2006Cat + PctUrbHi2006Cat,
         pctWetland2006 = PctWdWet2006Cat + PctHbWet2006Cat,
         areasqkm = round(areasqkm, 1),
         meandused = round(meandused, 1),
         inStreamCat = as.numeric(inStreamCat),
         sat = factor(sat, levels = c(5,7,8), labels = c('l5', 'l7', 'l8'))) 
  
refCorrect <- srMunged %>% select(UniqueID, blue, green, red, nir, sat) %>%
  gather(blue:nir, key = 'band' , value = 'value') %>%
  spread(sat, value) %>%
  group_by(band) %>%
  nest() %>%
  left_join(funcs.8) %>% #From 1_nhd_join_and_munge
  left_join(funcs.5) %>%
  mutate(pred8 = map2(lm8, data, predict),
         pred5 = map2(lm5, data, predict)) %>%
  select(-c(lm8,lm5)) %>%
  unnest(data, pred8, pred5) %>%
  select(-c(l8,l5)) %>%
  rename(l8 = pred8, l5 = pred5) %>% gather(l5,l7,l8, key = 'sat', value = 'value') %>%
  spread(band, value) %>%
  na.omit()
  
srMunged <- srMunged %>% select(-c(blue, green, red, nir, sat)) %>% left_join(refCorrect) %>%
  mutate(NR = nir/red,
         BG = blue/green,
         dWL = fui.hue(red, green, blue))

srMunged[srMunged == -9998] = NA ##Standardize missing values
##See if removing the ~2k obs with missing depth helps, it doesn't
#srMunged <- srMunged %>% filter(!is.na(meandused))

if(log == T){srMunged <- srMunged %>% mutate(value = log(value))}

## Select parameters to go into model
paste0(names(srMunged), collapse = ', ')

## These parameters make sense in terms of their mechanisms, resolution (with some assumptions), and reliability
features <- "NR, BG, dWL, inStreamCat, PctCarbResidCat, Precip8110Cat, Tmean8110Cat, RunoffCat, ClayCat, SandCat, OmCat, PermCat, RckdepCat, WtDepCat, BFICat, ElevCat, AOD, BFICat, KffactCat, CatAreaSqKm, PctAg2006Slp10Cat, ElevCat, NPDESDensCat, HydrlCondCat, PctImp2006Cat, pctUrban2006, pctForest2006, PctCrop2006Cat, pctWetland2006, WetIndexCat, areasqkm, meandused"

features <- str_split(features, pattern = ', ') %>% unlist()

##Pull holdout data
set.seed(340987)
holdOut <- srMunged %>%
  group_by(region) %>%
  sample_frac(.2)

df <- srMunged %>% filter(!UniqueID %in% holdOut$UniqueID)

#One hot encoding for xgboost
oneHot <- onehot(df %>% select(features))
df.encoded <- predict(oneHot, df %>% select(features)) %>% as.data.frame(.) 

# Split up cross validation folds based on stratified sampling
df <- df %>% mutate(SpaceTime = paste0(region, sat),
                    group1 = cut_number(lat, 3),
                    date = ymd(date),
                    julian = as.numeric(julian.Date(date))) %>%
            group_by(group1) %>%
            mutate(group2 = cut_number(long, 3)) %>%
            ungroup() %>%
            mutate(spaceCluster = paste0(group1,group2),
                   timeCluster = cut_number(julian, 9))

folds <- CreateSpacetimeFolds(df, spacevar = 'spaceCluster', timevar = 'timeCluster', k= 3, seed = 34985)

control <- trainControl(method="cv", savePredictions = 'none', 
                        returnResamp = 'final', index = folds$index, 
                        indexOut = folds$indexOut)

## Do initial feature selection with conservative hyperparameters
ffs.tuneGrid <- expand.grid(nrounds = 300, max_depth = 4, eta = .1, 
                            gamma = 1, min_child_weight = 1, colsample_bytree = .6,
                            subsample = .6)

# Takes awhile, so set it up all to run in parrallel
cl <- makePSOCKcluster(availableCores() - 4)
registerDoParallel(cl)

ffs <- ffs(df.encoded, df$value, method = 'xgbTree', metric = 'RMSE', 
           tuneGrid = ffs.tuneGrid, trControl = control, verbose = T)

stopCluster(cl)

ffsResults <- ffs$perf_all

# Save the results
write_feather(ffsResults, 'out/ffsResultsFull.feather')

ffsResults %>%
  group_by(nvar) %>%
  summarise(RMSE = median(RMSE),
            SE = median(SE)) %>%
  ggplot(.) + geom_line(aes(x = nvar, y = RMSE)) +
  geom_errorbar(aes(x = nvar, ymin = RMSE - SE, ymax = RMSE + SE), color = 'red')

#ggsave('figures/rfeRMSE.png', device = 'png', width = 6, height = 4, units = 'in')
```

### XGBoost with Default Hyperparameters

```{r baseline model, message=FALSE}
ffsResults <- read_feather('out/ffsResultsFull.feather')
features <- ffsResults[ffsResults$RMSE == min(ffsResults$RMSE),] %>%
  select(-c(nvar, RMSE, SE)) %>%
  paste(.) %>% .[.!= 'NA']

oneHot <- onehot(df %>% select(features))
df.encoded <- predict(oneHot, df %>% select(features)) %>% as.data.frame(.) 

# We'll set up a tuning grid with default values, define a `trainControl()`
# function (which in this case disables printing training log to console
# plus disables resampling as we want to use all training data to fit the
# model), and finally fit the model with `caret::train()`:

grid_default <- expand.grid(
  nrounds = 100,
  max_depth = 4,
  eta = 0.3,
  gamma = 0,
  colsample_bytree = .8,
  min_child_weight = .8,
  subsample = 1
)

#Set up a cluster to run everything in parrallel
cl <- makePSOCKcluster(availableCores()/2)
registerDoParallel(cl)

train_control <- caret::trainControl(method="cv", savePredictions = T, 
                          returnResamp = 'final', index = folds$index, 
                          indexOut = folds$indexOut)

xgb_base <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = train_control,
  tuneGrid = grid_default,
  method = "xgbTree",
  verbose = TRUE
)
```

# Grid Search for Hyperparameters

```{r First tune, warning=FALSE, fig.height=7}
# Next, we build tuning grid for `caret` to explore between different
# hyperparameter values. To get started we're using some suggestions from
# [here](https://www.slideshare.net/OwenZhang2/tips-for-data-science-competitions/14).
# 
# ## Step 1: Number of Iterations and the Learning Rate
# 
# We going to start the tuning "the bigger knobs" by setting up the maximum
# number of trees:

# note to start nrounds from 200, as smaller learning rates result in errors so
# big with lower starting points that they'll mess the scales


# To get reasonable running time while testing hyperparameter combinations
# with `caret` we don't want to go over `r nrounds`. Then, we want to find a good enough learning rate for this number of trees,
# as for lower learning rates `r nrounds` iterations might not be enough.
# 
# Next, as the maximum tree depth is also depending on the number of iterations
# and the learning rate, we want to experiment with it at this point to narrow
# down the possible hyperparameters. We'll also create a helper function to 
# create the visualizations with `ggplot2`, called `tuneplot()`:

nrounds <- 1000

tune_grid <- expand.grid(
  nrounds = seq(from = 100, to = nrounds, by = 100),
  eta = c(0.05, 0.1, 0.3),
  max_depth = c(2, 3, 4),
  gamma = 0,
  colsample_bytree = 1,
  min_child_weight = 1,
  subsample = 1
)

tune_control <- caret::trainControl(
  method = "cv", # cross-validation
  index = folds$index,
  indexOut = folds$indexOut,# with n folds 
  verboseIter = FALSE, # no training log
  allowParallel = TRUE # FALSE for reproducible results 
)

xgb_tune <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = tune_control,
  tuneGrid = tune_grid,
  method = "xgbTree",
  verbose = TRUE
)

# helper function for the plots
tuneplot <- function(x, probs = .90) {
  ggplot(x) +
    coord_cartesian(ylim = c(quantile(x$results$RMSE, probs = probs), min(x$results$RMSE))) +
    theme_bw()
}

tuneplot(xgb_tune)
xgb_tune$bestTune
```


```{r second tune, warning=FALSE, fig.height=5}

# With `r nrounds` iterations a learning rate of `r xgb_tune$bestTune$eta`
# seems to be a good starting point. The model currently has a cross-validated
# RMSE of `r round(min(xgb_tune$results$RMSE), digits = 5)`. Also to be noted 
# that with `max_depth = 2` and smaller learning rates the forest seems not yet
# stabile.
# 
# ## Step 2: Maximum Depth and Minimum Child Weight
# 
# After fixing the learning rate to `r xgb_tune$bestTune$eta` and we'll also
# set maximum depth to `r xgb_tune$bestTune$max_depth` +-1 (or +2 if 
# `max_depth == 2`) to experiment a bit around the suggested best tune in
# previous step. Then, well fix maximum depth and minimum child weight:


tune_grid2 <- expand.grid(
  nrounds = seq(from = 50, to = nrounds, by = 50),
  eta = xgb_tune$bestTune$eta,
  max_depth = ifelse(xgb_tune$bestTune$max_depth == 2,
    c(xgb_tune$bestTune$max_depth:4),
    xgb_tune$bestTune$max_depth - 1:xgb_tune$bestTune$max_depth + 1),
  gamma = 0,
  colsample_bytree = 1,
  min_child_weight = c(1, 2, 3),
  subsample = 1
)

xgb_tune2 <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = tune_control,
  tuneGrid = tune_grid2,
  method = "xgbTree",
  verbose = TRUE
)

tuneplot(xgb_tune2)
xgb_tune2$bestTune
```


```{r third tune, warning=FALSE, fig.height=5}
# The second tune results RMSE of `r round(min(xgb_tune2$results$RMSE), digits=5)`,
# `r glue("{round((min(xgb_tune$results$RMSE)/min(xgb_tune2$results$RMSE) - 1)*100, digits=2)}%")`
# over the previous step.
# 
# ## Step 3: Column and Row Sampling
# Based on this, we can fix minimum child weight to `r xgb_tune2$bestTune$min_child_weight`
# and maximum depth to `r xgb_tune2$bestTune$max_depth`. Next, we'll try
# different values for row and column sampling:

tune_grid3 <- expand.grid(
  nrounds = seq(from = 50, to = nrounds, by = 50),
  eta = xgb_tune$bestTune$eta,
  max_depth = xgb_tune2$bestTune$max_depth,
  gamma = 0,
  colsample_bytree = c(0.4, 0.6, 0.8, 1.0),
  min_child_weight = xgb_tune2$bestTune$min_child_weight,
  subsample = c(0.5, 0.75, 1.0)
)

xgb_tune3 <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = tune_control,
  tuneGrid = tune_grid3,
  method = "xgbTree",
  verbose = TRUE
)

tuneplot(xgb_tune3, probs = .95)
xgb_tune3$bestTune
```

```{r fourth tune, warning=FALSE, fig.height=5}

# Based on the above, we can set column sampling to `r xgb_tune3$bestTune$colsample_bytree`
# and row sampling to `r xgb_tune3$bestTune$subsample`. RMSE at this point is `r round(min(xgb_tune3$results$RMSE), digits=5)`, which is
# `r glue("{round((min(xgb_tune2$results$RMSE)/min(xgb_tune3$results$RMSE)-1) * 100, digits=2)}%")`
# better than RMSE in step 2.
# 
# ## Step 4: Gamma
# Next, we again pick the best values from previous step, and now will see 
# whether changing the gamma has any effect on the model fit:
# 


tune_grid4 <- expand.grid(
  nrounds = seq(from = 50, to = nrounds, by = 50),
  eta = xgb_tune$bestTune$eta,
  max_depth = xgb_tune2$bestTune$max_depth,
  gamma = c(0, 0.05, 0.1, 0.5, 0.7, 0.9, 1.0),
  colsample_bytree = xgb_tune3$bestTune$colsample_bytree,
  min_child_weight = xgb_tune2$bestTune$min_child_weight,
  subsample = xgb_tune3$bestTune$subsample
)

xgb_tune4 <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = tune_control,
  tuneGrid = tune_grid4,
  method = "xgbTree",
  verbose = TRUE
)

tuneplot(xgb_tune4)
xgb_tune4$bestTune
```


```{r fifth tune, warning=FALSE}
# After the inspection, we can set the gamma to `r xgb_tune4$bestTune$gamma`. 
# RMSE of this model is `r round(min(xgb_tune4$results$RMSE), digits=5)`
# `r if(xgb_tune4$bestTune$gamma!=0){glue("which is {round((min(xgb_tune3$results$RMSE)/min(xgb_tune4$results$RMSE)-1)*100, digits=2)}% improvement.")}else{"."}`
#
# ## Step 5: Reducing the Learning Rate
# 
# Now, we have tuned the hyperparameters and can start reducing the learning
# rate to get to the final model:

tune_grid5 <- expand.grid(
  nrounds = seq(from = 100, to = 1000, by = 50),
  eta = c(0.05, 0.1, .15), #c(0.01, 0.015, 0.025, 0.05, 0.1),
  max_depth = xgb_tune2$bestTune$max_depth,
  gamma = xgb_tune4$bestTune$gamma,
  colsample_bytree = xgb_tune3$bestTune$colsample_bytree,
  min_child_weight = xgb_tune2$bestTune$min_child_weight,
  subsample = xgb_tune3$bestTune$subsample
)

xgb_tune5 <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = tune_control,
  tuneGrid = tune_grid5,
  method = "xgbTree",
  verbose = TRUE
)

tuneplot(xgb_tune5)
xgb_tune5$bestTune
```


```{r final model, comment=NA}

# These will be the hyperparameters for the final model. RMSE now sets to 
# `r round(min(xgb_tune5$results$RMSE), digits=5)`, resulting in improvement of
# `r glue("{round((min(xgb_tune4$results$RMSE)/min(xgb_tune5$results$RMSE)-1)*100, digits=2)}%")`
# over the step 4, which in total is
# `r glue("{round((min(xgb_tune$results$RMSE)/min(xgb_tune5$results$RMSE)-1)*100, digits=2)}%")`
# better than the first tune.
# 
# ## Fitting the Model
# 
# Now that we have determined the parameters we want to use, we will use the
# training data (excluding the hold-out set which we will soon use to measure
# the model performance) without resampling to fit the model:


(final_grid <- expand.grid(
  nrounds = xgb_tune5$bestTune$nrounds,
  eta = xgb_tune5$bestTune$eta,
  max_depth = xgb_tune5$bestTune$max_depth,
  gamma = xgb_tune5$bestTune$gamma,
  colsample_bytree = xgb_tune5$bestTune$colsample_bytree,
  min_child_weight = xgb_tune5$bestTune$min_child_weight,
  subsample = xgb_tune5$bestTune$subsample
))

# cl <- makePSOCKcluster(6)
# registerDoParallel(cl)

(model <- caret::train(
  x = df.encoded,
  y = df$value,
  trControl = train_control,
  tuneGrid = final_grid,
  method = "xgbTree",
  verbose = TRUE
))

stopCluster(cl)
```


```{r hold-out RMSE}

# Note, that this model is not trained with whole training set yet which we will
# still do before submitting the predictions.
# 
# # Evaluating the Model Performance
# 
# Now we will evaluate the model performance by:
# 
#  - Testing the performance of the two baseline models and the final model with the hold-out data
#  - Submitting predictions to Kaggle from default hyperparameter XGBoost plus the tuned one 
#  - Testing whether taking the hold-out set had any effect on the cross-validated RMSE
#   
# ## With the Hold-out Set
# 
# By testing the performance with the hold-out set, we can see the effects that 
# the tuning had over the two baseline models: 

holdout_x <- predict(oneHot, holdOut %>% select(features))

holdout_y <- holdOut$value

if(log == T){
output <- tibble(Actual = exp(holdout_y), Predicted = exp(predict(model, holdout_x)), log = 'ln', UniqueID = holdOut$UniqueID)
}else{output <- tibble(Actual = holdout_y, Predicted = predict(model, holdout_x), log = 'none', UniqueID = holdOut$UniqueID)}

save(model, file = paste0('models/',iteration, '.Rdata'))
write_feather(output, paste0('out/outputs/',iteration, '.feather'))
```

```{r Eval}
output <- read_feather(paste0('out/outputs/',iteration, '.feather'))

evals <- output %>%
  summarise(rmse = rmse(Actual, Predicted),
            mae = mae(Actual, Predicted),
            mape = mape(Actual, Predicted),
            bias = bias(Actual, Predicted),
            p.bias = percent_bias(Actual, Predicted),
            smape = smape(Actual, Predicted)) %>%
  mutate(iteration = iteration)

evals %>% kable(digits = 2) %>% kable_styling() %>% scroll_box(width = '4in')

ggplot(output, aes(x = Actual +.5, y = Predicted +.5)) + 
  geom_hex(aes(fill = ..count..)) + 
  scale_fill_viridis(name = 'Point\nCount', trans = 'log10') + 
  geom_abline(color = 'red') + 
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  #scale_x_continuous(trans = 'log10', labels = scales::comma) +
  #scale_y_continuous(trans = 'log10', labels = scales::comma) +
  #coord_equal(ratio = 1) +
  labs(title = 'Hold-Out Validation', subtitle = 'Red line is 1:1', x = 'Actual', y = 'Predicted')

ggsave('figures/ModelValidation.png', width = 4, height = 3.5, units = 'in')

output.full <- output %>%
  left_join(srMunged, by = 'UniqueID') %>%
  mutate(residual = Actual - Predicted,
         year = year(date),
         month = month(date))

##look for errors associated with certain things
output.full %>% ggplot(., aes(x = residual, y = AOD)) + geom_hex()

## Some weird hacks for reordering factors
reorder_within <- function(x, by, within, fun = mean, sep = "___", ...) {
  new_x <- paste(x, within, sep = sep)
  stats::reorder(new_x, by, FUN = fun)
}

scale_x_reordered <- function(..., sep = "___") {
  reg <- paste0(sep, ".+$")
  ggplot2::scale_x_discrete(labels = function(x) gsub(reg, "", x), ...)
}

## Summary metrics across space time and observed value
errorSum <- output.full %>%
  #filter(Param == 'secchi') %>%
  mutate(Observed.Value = Actual) %>%
  rename(Year = year, Latitude = lat, Longitude = long) %>%
  gather(Observed.Value, Year, Latitude, Longitude, key = 'Variable', value = 'Value') %>%
  group_by(Variable, log) %>%
  mutate(quantile = cut_number(Value, 10, right = F, labels = F),
         quantLabs = cut_number(Value, 10,  right = F, dig.lab = 1)) %>%
  ungroup() %>%
  group_by(quantile, quantLabs, Variable, log) %>%
  dplyr::summarise(mae = mae(Actual, Predicted),
            smape = smape(Actual, Predicted),
            p.bias = percent_bias(Actual, Predicted),
            #mae = mae(Actual, Predicted),
            bias = bias(Actual, Predicted)) %>%
  gather(mae:bias, key = 'Metric', value = 'Error') %>%
  as.data.frame() %>%
  arrange(Variable, quantile) %>%
  mutate(order = row_number())

spaceTimeFigs()
spaceTimeFigs(abs = F)

output.full %>% mutate(yearGroups = cut_number(year, 20)) %>%
  ggplot(., aes(x = yearGroups, y = residual)) + 
  #geom_boxplot() +
  geom_violin(draw_quantiles = c(.25,.5,.75)) + 
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
  
  group_by(yearGroups) %>%
  summarise(mean.residual = mean(residual),
            sd.residual = sd(residual)) %>%
  ggplot(., aes(x = yearGroups, y = mean.residual)) + geom_point() + geom_errorbar(aes(ymin = mean.residual - sd.residual, ymax = mean.residual + sd.residual)) + theme(axis.text.x = element_text(angle = 90, vjust = .5))
  
##View the feature importance
model$modelInfo$varImp(model$finalModel) %>%
  mutate(Feature = fct_reorder(rownames(.), Overall, .desc = T)) %>%
  arrange(Overall) %>%
  ggplot(., aes(x = Feature, y = Overall)) + geom_col() + coord_flip() +
  ggtitle('Feature Importance')

ggsave('figures/FeatImp.png', width = 3.5, height = 3, units = 'in')
```



```{r, fig.height= 8}
output.full %>% 
  group_by(sat) %>%
  summarise(mae = mae(Actual, Predicted),
            smape = smape(Actual, Predicted),
            p.bias = percent_bias(Actual, Predicted),
            #mae = mae(Actual, Predicted),
            bias = bias(Actual, Predicted)) %>%
  gather(mae:bias, key = 'Metric', value = 'Error') %>%
  ggplot(., aes(x = Metric, y = Error, fill = sat)) +
  geom_col(position = 'dodge') +
  theme_bw() + 
  ggtitle('Absolute and Relative Error by Satellite')
  
val.conf <- output.full %>%
  mutate(smape = 2*(abs(Actual - Predicted)/(abs(Actual) + abs(Predicted))),
         p.bias = (Actual - Predicted)/abs(Actual)) %>%
  group_by(region, sat, log) %>%
  summarise(smape = sd(smape),
            p.bias = sd(p.bias)) %>%
  mutate(p.bias = ifelse(is.na(p.bias), 0, p.bias),
         smape = ifelse(is.na(smape), 0, smape)) %>%
  gather(smape, p.bias, key = 'Metric', value = 'sd')
  
output.full %>%
  group_by(region, sat, log) %>%
  summarise(smape = smape(Actual, Predicted),
            p.bias = percent_bias(Actual, Predicted),
            count = n()) %>%
  gather(smape:p.bias, key = 'Metric', value = 'value') %>%
  left_join(val.conf %>% select(region, sat, Metric, sd)) %>%
  ggplot(.) +
  scale_fill_viridis_c(trans = 'log10') +
  #geom_text(x = .5, y = .5) +
  geom_col(position = 'dodge', aes(x = Metric, y = value, fill = count, color = sat, group = sat)) +
  geom_errorbar(position = position_dodge(.9), aes(x = Metric, ymin = value - sd, ymax = value + sd, group = sat), width = .2) +
  #scale_y_continuous(labels = scales::percent) +
  theme(legend.position = 'top') +
  ggtitle('Error by region') +
  facet_wrap(~region, scales = 'free')#, ncol = 4)

ggplot(output.full, aes(x = residual, color = sat)) + geom_density() + facet_wrap(~region)


ggplot(output.full, aes(x = Actual, y = Predicted)) + 
  geom_point(alpha = .3) +
  geom_smooth(method = 'lm', aes(color = sat)) +
  #scale_fill_viridis(name = 'Point\nCount', trans = 'log10') + 
  #geom_abline(color = 'red') + 
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  #scale_x_continuous(trans = 'log10', labels = scales::comma) +
  #scale_y_continuous(trans = 'log10', labels = scales::comma) +
  #coord_equal(ratio = 1) +
  facet_wrap(~region, scales = 'free', shrink = T) +
  labs(title = 'Regional Evaluation Performance') + theme_bw()

ggsave('figures/RegionalEval.png', width = 6.5, height =6, units = 'in')

dummy <- expand.grid(year = seq(1985,2018), month = c(1:12), region = factor(unique(region$region)))

fieldmatch.coin <- output.full %>%
  filter(year < 2019,
         month %in% c(5:9)) %>%
  rename(Modelled = Predicted, In.Situ = Actual) %>%
  gather(Modelled, In.Situ, key = 'Measure', value = 'value') %>%
  group_by(year, COMID, Measure, region) %>%
  summarise(value = median(value)) %>%
  group_by(year, Measure, region) %>%
  summarise(mean = mean(value))


ggplot(fieldmatch.coin, aes(x = year, y = mean, color = Measure)) + 
  #geom_point(alpha = .1) +
  geom_line() + 
  scale_color_viridis_d(end = .6) +
  facet_wrap(~region, scales = 'free', ncol = 2) +
  labs(title = 'Smoothed SDD Trends By Region Using Coincident Testing Data') +
  theme_bw() +
  theme(legend.position = 'bottom')


## Next do it with all the field data we got.  
# We add the additional step of averaging per lake before per month because sampling is really uneven. 
# Load in lagos and wqp data
# sites <- read_feather('../Aquasat/2_rsdata/out/unique_site_inventory.feather')
# 
# wqp.lagos <- read_feather('../Aquasat/1_wqdata/out/wqp_lagos_unity.feather') %>%
#   dplyr::select(SiteID, date_unity, source, value = secchi) %>%
#   na.omit() %>%
#   filter(value > 0.01, 
#          value < 15) %>%
#   left_join(sites) %>%
#   filter(!is.na(lat)) %>%
#   mutate(year = year(date_unity),
#          month = month(date_unity)) %>%
#   filter(month %in% c(5:9),
#          year > 1983, year < 2019) %>%
#   group_by(year, SiteID, lat, long) %>%
#   dplyr::summarize(median = median(value),
#                    count = n())
# 
# plan(multiprocess)
# site.join <- wqp.lagos %>%
#   ungroup() %>%
#   distinct(SiteID, lat, long) %>%
#   st_as_sf(coords = c('long', 'lat'), crs = 4326) %>%
#   st_transform(st_crs(region)) %>%
#   split(., c(1:10)) %>%
#   future_map(~st_join(.,region, join = st_nearest_feature), .progress = T)
# plan(sequential)
# 
# site.join <- rbind(site.join[[1]], site.join[[2]], site.join[[3]], site.join[[4]], site.join[[5]],
#                     site.join[[6]], site.join[[7]], site.join[[8]], site.join[[9]], site.join[[10]])
# 
# fieldMean <- wqp.lagos %>%
#   left_join(site.join) %>%
#   group_by(year, region) %>%
#   dplyr::summarize(mean = mean(median, na.rm = T),
#             sd = sd(median, na.rm =T),
#             count.obs = sum(count),
#             count.sites = n())
# 
# write_feather(fieldMean, 'out/RegionalsInSituMeans.feather')

## A little backwords but read in the bootstrapped predictions from 5_lake_analysis
fieldmatch.full <- read_feather('out/RegionalsInSituMeans.feather') %>% mutate(Measure = 'In.Situ') %>%
  select(year, region, mean, se = sd, Measure) %>% bind_rows(read_feather('out/TS_Preds/NLA_bootstrapped.feather') %>% mutate(Measure = 'Modelled'))


p1 <- fieldmatch.coin %>%
  ggplot(., aes(x = year, y = mean, color = Measure)) + 
  geom_line() + 
  scale_color_viridis_d(end = .6) +
  facet_wrap(~region, scales = 'free_y', ncol = 1) +
  labs(title = 'Coincident\nObservations') +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_blank())

p2 <- fieldmatch.full %>%
  ggplot(., aes(x = year, y = mean, color = Measure)) + 
  geom_line() + 
  scale_color_viridis_d(end = .6) +
  facet_wrap(~region, scales = 'free_y', ncol = 1) +
  labs(title = 'All\nObservations') +
  theme_bw() +
  theme(legend.position = 'right',
        axis.title = element_blank())

g <- grid.arrange(p1,p2, ncol = 2, widths = c(1,1.7), left = 'Mean Clarity (m)', bottom = 'Year')


ggsave('figures/CoinVFullObs.png', plot = g, width = 5.5, height = 9, units = 'in')


##Fig for seminar pres
fieldmatch.full %>% #filter(Measure == 'In.Situ') %>%
  ggplot(., aes(x = year, y = mean, color = Measure)) + 
  geom_line() + 
  scale_color_viridis_d(end = .6) +
  facet_wrap(~region, scales = 'free_y', ncol = 2) +
  labs(title = 'Full Field and Satellite Observations') +
  theme_bw() +
  theme(legend.position = c(.75,.1),
        axis.title = element_blank())

ggsave('figures/Full_TS_field.png', width = 7, height = 6, units = 'in')
```
