==> 7_lake_level_analysis.Rmd <==
---
title: "7_lake_level_analysis"
author: "Simon Topp"
date: "4/20/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---
## Individual lake analysis

```{r}
## Pull in original predictions for each lake
summ.meds.full <- read_feather(paste0('out/TS_Preds/NLA2012_cntr_',iteration,'.feather')) %>% mutate(sample = 'NLA') %>% #filter(COMID %in% nla.2012$COMID) %>%
  bind_rows(read_feather(paste0('out/TS_Preds/Over10_',iteration,'.feather')) %>% mutate(sample = 'Over10')) %>%
  bind_rows(read_feather(paste0('out/TS_Preds/EcoReg2000_cntr_',iteration,'.feather')) %>%
              mutate(sample = 'Random')) %>%
  filter(month %in% c(5:9),
  year < 2019) %>% ##Incomplete obs for 2019 so we won't use it.
  group_by(COMID, year, region) %>%
  summarise(secchi.med = median(value))

#Pull in all the lake metadata
lake.join <- read_feather('out/NLA2012LakesFull.feather') %>% mutate(sample = 'NLA') %>%
  bind_rows(read_feather('out/Over10LakesFull.feather') %>% mutate(sample = 'Over10')) %>%
  bind_rows(read_feather('out/EcoReg2000LakesFull.feather') %>% mutate(sample = 'Random')) %>%
  distinct(COMID, .keep_all = T)

##Real Quick look at differences between the samples
lake.join %>% filter(sample == 'NLA'|sample == 'Random', COMID %in% summ.meds.full$COMID) %>%
  ggplot() + geom_density(aes(x = areasqkm, color = sample)) + scale_x_log10() + facet_wrap(~region)


check <- lake.join %>% filter(sample == 'NLA'|sample == 'Random', COMID %in% summ.meds.full$COMID) %>%
  group_by(sample, region) %>% summarise(meanSize = median(areasqkm)) %>% spread(sample, meanSize) %>% mutate(dif = NLA/Random)

#For individual lake analysis use stricter filter for total years with observations
counts <- summ.meds.full %>%
  group_by(COMID, region) %>%
  summarise(count = n()) %>%
  filter(count > 25)

summ.meds.full <- summ.meds.full %>% filter(COMID %in% counts$COMID)

##Calculate mann-Kendall statistics for each lake that has over 25 years of observations
mk <- summ.meds.full %>% 
  group_by(COMID, region) %>%
  arrange(year) %>%
  nest() %>%
  mutate(mk = purrr::map(data, ~sens.slope(.$secchi.med)),
         sen.slope = purrr::map_dbl(mk, 'estimates'),
         sen.slope = sen.slope*100,
         p.value = purrr::map_dbl(mk, 'p.value'),
         sig = ifelse(p.value < .1, T, F)) %>%
  select(COMID, region, sen.slope, p.value, sig) %>%
  left_join(lake.join) %>%
  mutate(Size.Group = cut(areasqkm, breaks = c(0,1,10,100,2000), labels = c('<1','1-10','10-100','>100'))) %>%
  filter(!is.na(Size.Group))

#write_feather(mk, 'out/LakeLevelMK.feather' )
```


```{r}
mk <- read_feather('out/LakeLevelMK.feather')

# Get median trend by size class.
mk.ave <- mk %>% group_by(Size.Group) %>% summarise(mean =mean(sen.slope), median = median(sen.slope), sd = sd(sen.slope), lower90 = mean - sd, upper90 = mean + sd, count = n())

mk %>% group_by(Size.Group) %>% summarise(mean =mean(sen.slope), median = median(sen.slope))

# Coarse above/below 10 values
mk %>% mutate(b10 = ifelse(areasqkm < 10, T, F)) %>% group_by(b10) %>% summarise(mean =mean(sen.slope), median = median(sen.slope), sd = sd(sen.slope), lower90 = mean - sd, upper90 = mean + sd, count = n())

ggplot(mk, aes(x = sen.slope)) + geom_density(aes(fill = Size.Group), alpha = .4) + scale_fill_viridis_d(end = .7) + theme_bw() + labs(title = 'Distribution of Lake Trends by Size Class', x = 'Thiel-Sen Slope (cm/yr)', y = 'Density', color = 'Size Group\n(sq km)', fill = 'Size Group\n(sq km)') + geom_vline(data = mk.ave, aes(xintercept = median, color = Size.Group)) + scale_color_viridis_d(end = .9) + coord_cartesian(xlim = c(-5,5))

ggsave('figures/TrendbySizeClass.png', width = 6, height = 3, units = 'in')

## Try another version
p1 <- ggplot(mk %>% left_join(mk.ave), aes(x = Size.Group, y = sen.slope, group = Size.Group))  +
  geom_violin() + 
  geom_boxplot(width = .1, fill = 'grey60') +
  geom_hline(yintercept = 0, color = 'red') +
  #scale_fill_viridis_c('Number of\nLakes', option = 'plasma', end = .8, trans = 'log10') +
  coord_cartesian(ylim = c(-5,5))+
  labs(x = 'Size Group (sq. km)', y = 'Slope Distribution (cm/year)') +
  theme_bw()

ggsave('figures/TrendbySizeClassViolinV2.png', width = 2.5, height = 3, units = 'in')

#Test for significance between size groups
wilcox.test(mk$sen.slope[mk$Size.Group == '<1'], mk$sen.slope[mk$Size.Group == '1-10'], conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '1-10'], mk$sen.slope[mk$Size.Group == '10-100'], conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '<1'|mk$Size.Group == '1-10'], mk$sen.slope[mk$Size.Group == '>100' |mk$Size.Group == '10-100'], conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '>100'], conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '10-100'],conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '1-10'],conf.int = T)
wilcox.test(mk$sen.slope[mk$Size.Group == '<1'],conf.int = T)

cor.test(mk$sen.slope,mk$areasqkm)


#Look at the distribution of clarity change across in-network and out-of-network lakes, no
ggplot(mk, aes(x = sen.slope)) + geom_density(alpha = .4, aes(fill = factor(inStreamCat)))

# Look at reservoirs vs natural lakes
ggplot(mk %>% left_join(mk.ave), aes(x = ftype, y = sen.slope, group = ftype))  +
  geom_violin() + 
  geom_boxplot(width = .1, fill = 'grey60') +
  geom_hline(yintercept = 0, color = 'red') +
  #scale_fill_viridis_c('Number of\nLakes', option = 'plasma', end = .8, trans = 'log10') +
  coord_cartesian(ylim = c(-5,5))+
  labs(x = 'Lake Type', y = 'Slope Distribution (cm/year)', title = 'Slope Distribution by lake Type') +
  theme_bw()
wilcox.test(mk$sen.slope[mk$ftype == 'LakePond'], mk$sen.slope[mk$ftype == 'Reservoir'], conf.int = T)

## Look at res breakdown by region
ggplot(mk, aes(x = ftype)) + geom_bar() + facet_wrap(~region) + scale_y_log10()

mk %>% group_by(inStreamCat) %>% summarise(mean = mean(sen.slope), med = median(sen.slope))
wilcox.test(mk$sen.slope[mk$inStreamCat == 0], mk$sen.slope[mk$inStreamCat == 1], conf.int = T)

##Quick look at size distributions from LAGOS to make the point that we
## generally sample larger lakes.
lagos.lakes <- read_csv('in/LAGOS/Lagos_lakeInfo.txt')
lagos.secchi <- readRDS('in/LAGOS/Sol_data.rds') %>%
  select(secchi, lagoslakeid, sampledate) %>% na.omit() %>%
  left_join(lagos.lakes) %>%
  mutate(areasqkm = lake_area_ha/100)

ggplot(lagos.secchi, aes(x = lake_area_ha)) + geom_density() + scale_x_log10()

mean(lagos.secchi$areasqkm)
median(lagos.secchi$areasqkm)

## Quick look at population density to following up on Stets et al 2020.
## Look at all lakes following up on Stets et al 2020
#Coarse
pDense <- mk %>% 
  select(sen.slope, region, PopDen2010Cat) %>%
  group_by(region) %>%
  mutate(PopDensity2010 = cut(PopDen2010Cat, quantile(PopDen2010Cat, c(0,.6,1)), labels = c('Lower', 'Upper'), T),
         p.value = wilcox.test(sen.slope[PopDensity2010 == 'Lower'], sen.slope[PopDensity2010 == 'Upper'])$p.value,
         SigDif = ifelse(p.value < .05, T, F)) %>%
  ungroup()

ggplot(check, aes(x = PopDensity2010, y = sen.slope, fill = SigDif)) + geom_boxplot() + facet_wrap(~region, scales = 'free') + labs(x = 'Population Density split on 50th Percentile', y = 'Clarity Trend', title = 'All Trends (n = 8,310)')


# Finer grain
# Make a figure showing the relationship between population density and observed trend
popGroups <- mk %>% mutate(pDens = cut_number(PopDen2010Cat, 5, labels = c('[0-20]','(20-40]', '(40-60]', '(60-80]','(80-100]')))

p2 <- ggplot(popGroups, aes(x = pDens, y = sen.slope))  +
  geom_violin() + 
  geom_boxplot(width = .1, fill = 'grey60') +
  geom_hline(yintercept = 0, color = 'red') +
  #scale_fill_viridis_c('Number of\nLakes', option = 'plasma', end = .8, trans = 'log10') +
  coord_cartesian(ylim = c(-5,5))+
  labs(x = 'Population Density Quantile', y = 'Slope Distribution (cm/year)') +
  theme_bw() + 
  theme(#axis.text.y = element_blank(),
        axis.title.y = element_blank())

ggsave('figures/TrendbyPopDense.png', width = 2.75, height = 3, units = 'in')

## Make combine popdense and lake size for figure 3
g <- grid.arrange(p1,p2, nrow = 1)
ggsave('figures/PopSizeViolins.png', plot = g, width = 5.5, height = 3.5, units = 'in')

popGroups %>% group_by(pDens) %>% summarise(median = median(sen.slope))

# Test for significance
wilcox.test(popGroups$sen.slope[popGroups$pDens == '[0-20]'], popGroups$sen.slope[popGroups$pDens == '(80-100]'])
wilcox.test(popGroups$sen.slope[popGroups$pDens == '[0-20]'], conf.int = T)


##Look at population density vs lakesize to make sure we're not just showing two correlated things.
ggplot(mk, aes(x = PopDen2010Cat, y = areasqkm)) + geom_point() +
  scale_x_continuous(trans = 'log10') +
  scale_y_continuous(trans = 'log10') +
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3)

ggplot(popGroups, aes(x = pDens, y = areasqkm))  +
  geom_violin() + 
  geom_boxplot(width = .1, fill = 'grey60') +
  geom_hline(yintercept = 0, color = 'red') +
  scale_y_continuous(trans = 'log10') +
  #labs(x = 'Population Density Quantile', y = 'Slope Distribution (cm/year)') +
  theme_bw()

cor.test(mk$PopDen2010Cat, mk$areasqkm)
summary(lm(areasqkm~PopDen2010Cat, data = mk))
## Just plot up the significant trends of individual lakes to look at nation wide patterns
ggplot() +
  geom_sf(data = region %>% st_simplify(dTolerance = 1000)) +
  geom_sf(data = mk %>%
    filter(sig == T) %>%
    st_as_sf(., coords = c('long', 'lat'), crs = 4326),
    aes(color = sen.slope)) +
  scale_color_gradient2(low = 'red', mid = 'grey60', high = 'blue') 

ggplot(mk, aes(x = areasqkm, y = sen.slope)) + geom_point() + scale_x_log10()
```
