==> 99__MnComps.Rmd <==
---
title: "99_MnComps"
author: "Simon Topp"
date: "3/27/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Set up shapefile to pull MN Lakes

```{r}
# Mess with NHD data
mnNHDPoly <- st_read('in/MN_LB_Comparison/NHDPlusSR/NHDPLus09/NHDSnapshot/Hydrography/NHDWaterbody.shp') %>% 
  rbind(st_read('in/MN_LB_Comparison/NHDPlusMS 2/NHDPLus10U/NHDSnapshot/Hydrography/NHDWaterbody.shp')) %>%
  rbind(st_read('in/MN_LB_Comparison/NHDPlusMS/NHDPLus07/NHDSnapshot/Hydrography/NHDWaterbody.shp')) %>% 
  rbind(st_read('in/MN_LB_Comparison/NHDPlusGL/NHDPLus04/NHDSnapshot/Hydrography/NHDWaterbody.shp'))

## Join NHD Lakes with lakebrowser lakes.
mnNHD <- mnNHDPoly %>%
  st_transform(4326) %>% st_centroid() %>% st_join(ecoregs, join = st_within) %>%
  filter(!is.na(US_L3NAME))
  filter(FTYPE == 'LakePond' | FTYPE == 'Reservoir',
         AREASQKM >= .01, ##Make sure they're landsat visible, even with this some aren't
         AREASQKM < 4000) ##Exclude the Great lakes

LBPreds <- st_read('in/MN_LB_Comparison/lakesmnwq_1975_2008_fill1975-epsg4326/MNWQ_1975_2008_fill1975-epsg4326.shp')

plan(multiprocess)
centerJoin <- mnNHD %>%
  split(., c(1:10)) %>%
  future_map(~st_join(.,LBPreds), .progress = T)
plan(sequential)

centerJoin <- rbind(centerJoin[[1]], centerJoin[[2]], centerJoin[[3]], centerJoin[[4]], centerJoin[[5]],
                    centerJoin[[6]], centerJoin[[7]], centerJoin[[8]], centerJoin[[9]], centerJoin[[10]]) %>%
  filter(!is.na(unmlknum))

### Filter MNPoly to those that overlap with LakeBrowser
mnOut <- mnNHDPoly %>% filter(COMID %in% centerJoin$COMID) %>% st_zm()

mnOut %>% group_by(FTYPE) %>% summarise(count = n())

## For some reason regional NHD data doesn't have the full suite of attributes
nhdOut <- st_read('D:/GIS_Data/NHD/nhdPlusV2_WaterBodiesFull.shp') %>%
  filter(COMID %in% mnOut$COMID) %>%  
           rename_at(vars(FDATE:LakeArea), tolower) %>%
  mutate_at(vars(shape_leng, shape_area, meandepth, lakearea), ~format(round(., 3), nsmall = 3)) %>%
  mutate(lakevolume = format(round(lakevolume, 2), nsmall = 2),
         COMID = factor(COMID))

#st_write(nhdOut, 'in/MN_LB_Comparison/LB_NHDLakes.shp')
nhdOut <- st_read('in/MN_LB_Comparison/LB_NHDLakes.shp')
```

## Bring in the MN data
```{r}
## Breakdown of MN to look compare against LakeBrowser
LBPreds <- st_read('in/MN_LB_Comparison/lakesmnwq_1975_2008_fill1975-epsg4326/MNWQ_1975_2008_fill1975-epsg4326.shp')

ecoregs <- st_read('in/MN_LB_Comparison/mn_eco_l3/mn_eco_l3.shp') %>%
  st_transform(4326)

LBPreds <- LBPreds %>% select(grep('MEAN', names(LBPreds), value = T)) %>%
  st_centroid() %>%
  st_join(ecoregs, join = st_nearest_feature)
LBPreds <- LBPreds %>% st_set_geometry(NULL)
LBPreds[LBPreds == 0] = NA

LBEcoMeans <-LBPreds %>%
  select(X1975MEAN:X2008MEAN, US_L3NAME) %>%
  gather(X1975MEAN:X2008MEAN, key = 'year', value = 'value') %>%
  filter(value != 0) %>%
  mutate(year = ifelse(year == 'X1975MEAN', 1975,
                       ifelse(year == 'X1985MEAN', 1985,
                              ifelse(year == 'X1990MEAN', 1990,
                                     ifelse(year == 'X1995MEAN', 1995,
                                            ifelse(year == 'X2000MEAN', 2000,
                                                   ifelse(year == 'X2005MEAN', 2005,
                                                          ifelse(year == 'X2008MEAN', 2008,NA))))))), 
         source = 'LB') %>%
  group_by(US_L3NAME, year, source) %>%
  summarise(value = mean(value))


#Compare GIS layer with manually extracted data
mnCheck <- read_csv('in/MN_LB_Comparison/LB_EcoAverages.csv') %>%
  select(Ecoregion, year = Year, value = Clarity.Mean) %>% mutate(source = 'manual') %>%
  bind_rows(LBEcoMeans %>% select(Ecoregion = US_L3NAME, year, value) %>%
              mutate(source = 'data'))

ggplot(mnCheck, aes(x = year, y = value, color = source)) + geom_line() + facet_wrap(~Ecoregion) 
```

## Set up prediction framework

```{r}
# Load in lakeCat data
lc.files <- list.files('in/lakeCat/unzip', full.names = T)

# Join lakes to nhd plus LakeCat data
for(i in lc.files){
  if(i == first(lc.files)){
    lc <- read.csv(i) %>% mutate(COMID = factor(COMID))
    lake.join.mn <- nhdOut %>% st_set_geometry(NULL) %>%
      mutate(COMID = factor(COMID)) %>%
      left_join(lc , by = 'COMID')
    }else{
    lc <- read.csv(i) %>% mutate(COMID = factor(COMID))%>%
      select(-c(CatAreaSqKm, WsAreaSqKm, CatPctFull,WsPctFull,inStreamCat))
    lake.join.mn <- lake.join.mn %>%
      left_join(lc, by = 'COMID')}
}

names.in <- names(lake.join.mn)
## Remove unwanted parameters and munge a couple of date related ones.
lake.join.mn <- lake.join.mn %>%
  filter(!is.na(CatAreaSqKm)) %>%  ## 
  select(-c(grep(names.in, pattern = '*Ws', value = T), 
            grep(names.in, pattern = '2011Cat', value = T),
            ## 2011 values aren't representative of the study period
            grep(names.in, pattern = 'PctFire', value = T), 
            ## While potentially important, the forest metrics lack 
            grep(names.in, pattern = 'PctFrstLoss', value = T), 
            ##the temporal resolution we need.
            grep(names.in, pattern = 'NABD', value = T), 
            ##Dam info acts as identifiers and are redundant with lake_type/area
            grep(names.in, pattern = 'Dam', value = T), 
            ##Dam info acts as identifiers and are redundant with lake_type/area
            'Precip08Cat', 'Precip09Cat', 'Tmean08Cat', 'Tmean09Cat'))
            #Use long term averages not these
#Round the majority of values to the nearest 1 to avoid variables becoming 'unique ID's)
round1 <- names(lake.join.mn %>% select(c(CatAreaSqKm:CatPctFull,PctImp2006Cat, PctCarbResidCat:WetIndexCat)))

## Round a few to 10 because they're variation is real big and larger values with 
#Round a subset to the nearest 10th because their values more or less range from 0 to 1.
round.1 <- names(lake.join.mn %>% select(c(AgKffactCat, KffactCat, MineDensCat)))

lake.join.mn <- lake.join.mn %>%
  mutate_at(round1, round, digits = 0) %>%
  mutate_at(round.1, round, digits = 1)

lake.join.mn[lake.join.mn == -9998] = NA

#write_feather(lake.join.mn, 'out/MN_LB_Comparison/lakeJoinMn.feather')
lake.join.mn <- read_feather('out/MN_LB_Comparison/lakeJoinMn.feather')

ffsVariables <- read_feather('out/ffsResultsFull.feather')
features <- ffsVariables[ffsVariables$RMSE == min(ffsVariables$RMSE),] %>%
  select(-c(nvar, RMSE, SE)) %>%
  paste(.) %>% .[.!= 'NA']

load('models/FFS_PolyCorr_noLn_Secchi.Rdata')


ids <- list.files('lake_data/LB_Lakes') %>% strsplit(split = '.csv', fixed = T) %>% unlist()
#ids <- ids[!ids %in% predicted$COMID]
paths <- list.files('lake_data/LB_Lakes', full.names = T)

munger <- function(id){
  df <- read.csv(grep(paste0('/',id,'.csv'),paths, value = T), stringsAsFactors = F) %>%
  mutate(COMID = as.character(COMID),
         year = year(date),
         UniqueID = row_number(),
         sat = factor(sat, levels = c(5,7,8), labels = c('l5','l7','l8'))) %>%
  filter(!is.na(blue),
         dswe == 1,
         dswe_sd < .4,
         cScore == 0)
  return(df)
}

SR.mn <- ids %>% map_dfr(munger)

df <- SR.mn %>%
  gather(blue, green, red, nir, key = 'band' , value = 'value') %>%
  spread(sat, value) %>%
  group_by(band) %>%
  nest() %>%
  #left_join(funcs.8) %>% #From 1_nhd_join_and_munge
  left_join(funcs.5) %>%
  mutate(#pred8 = map2(lm8, data, predict),
         pred5 = map2(lm5, data, predict)) %>%
  select(-lm5) %>%
  unnest(c(data, pred5)) %>%
  select(-l5) %>%
  rename(l5 = pred5) %>% gather(l5,l7, key = 'sat', value = 'value') %>%
  spread(band, value) %>%
  filter(!is.na(blue)) %>%
  inner_join(lake.join.mn %>% mutate(COMID = as.character(COMID))) %>%
  distinct(.keep_all = T) %>%
  mutate(NR = nir/red,
         BG = blue/green,
         dWL = fui.hue(red, green, blue),
         #date = ymd_hms(date),
         month = as.numeric(month(date)),
         pctForest2006 = PctDecid2006Cat + PctConif2006Cat + PctMxFst2006Cat,
         pctUrban2006 = PctUrbMd2006Cat + PctUrbHi2006Cat,
         pctWetland2006 = PctWdWet2006Cat + PctHbWet2006Cat,
         areasqkm = round(areasqkm, 1),
         meandused = round(meandused, 1)) %>%
  filter(pixelCount > 5) %>%
  filter_at(vars(blue,green,red,nir,swir1,swir2),all_vars(.>0 & .< 2000))

encoder <- onehot(df %>% select(features))
lake.input <- predict(encoder, df %>% select(features))

value = tibble(value = predict(model, lake.input))

predicted <- df %>%
  select(system.index, COMID, date, year, month) %>%
  bind_cols(value) %>%
  left_join(mnNHD %>% select(COMID, Ecoregion = US_L3NAME) %>% mutate(COMID = factor(COMID)))

#write_feather(predicted %>% select(-geometry), 'out/MN_LB_Comparison/srPreds.feather')
predicted <- read_feather('out/MN_LB_Comparison/srPreds.feather')
## Take another look at how they match up 

predsMN <- predicted %>%
  filter(year %in% c(1985,1990,1995,2000,2005, 2008),
         month %in% c(8,9)) %>%
  group_by(Ecoregion, year, COMID) %>%
  summarise(value = mean(value, na.rm = T)) %>%
  group_by(Ecoregion, year) %>%
  summarise(value = mean(value)) %>%
  mutate(source = 'RS') %>%
  bind_rows(mnCheck %>% filter(year %in% c(1985,1990,1995, 2000,2005, 2008)))

p1 <- ggplot(predsMN %>% filter(source != 'manual'), aes(x = year, y = value, color = source)) + 
  geom_line() + geom_point(alpha = .4) +
  theme_bw() +
  scale_color_viridis_d(end = .8, labels = c('LakeBrowser', 'This Study')) +
  facet_wrap(~Ecoregion, nrow = 2) +
  theme(legend.position = c(.9,.2),
        axis.text.x = element_text(angle = 30, hjust = 1)) +
  labs(x = 'Year', y = 'Predicted Secchi (m)', title = 'Lake Browser Timeseries Comparisons')

ggsave('figures/__MN_TS_Comps.png', width = 5, height = 4, units = 'in')

predsMN %>%
  spread(source, value) %>%
  mutate(Driftless = ifelse(Ecoregion == 'Driftless Area', 'Driftless Area', 'Other Regions'))  %>%
  ggplot(.,aes(x = RS, y = data)) + geom_point(aes(color = Driftless)) +
  geom_smooth(method = 'lm', se = F) +
  #geom_abline(color = 'red') + 
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  ggpubr::stat_regline_equation(vjust = 2.4) +
  labs(x = 'This Study', y = 'LakeBrowser Predictions', color = '') + 
  theme_bw()  

ggsave('figures/__MN_OverallComps.png', width = 4, height = 3, units = 'in')

lms <- predsMN %>%
  group_by(source, Ecoregion) %>%
  nest() %>%
  mutate(lm = purrr::map(data, ~lm(value~year,.)),
         tidy = purrr::map(lm, broom::tidy)) %>%
  unnest(tidy) %>%
  select(-lm, -data) %>%
  filter(term == 'year') %>%
  mutate(Significant = ifelse(p.value < 0.05, T,F))

ggplot(lms %>% filter(source %in% c('RS', 'manual')), aes(x = Ecoregion, y = estimate, color = source)) + 
  geom_point() +
  geom_errorbar(aes(ymin = estimate - 2*std.error, ymax = estimate + 2*std.error)) +
  theme_bw() +
 theme(axis.text.x = element_text(angle = 20, hjust = 1)) + 
  scale_color_discrete('Source', labels = c('LB_Shapefile', 'Our Preds')) +
  labs(title = 'Comparison of overall trends', x = 'Ecoregion', y = 'Linear Trend (m/y, 95% CI)')

ggsave('figures/__MN_TrendComps.png', width = 4.5, height = 3, units = 'in')


mnWide <- predsMN %>%
  spread(source, value)
bias(mnWide$manual, mnWide$RS)
rmse(mnWide$manual, mnWide$RS)
mae(mnWide$manual, mnWide$RS)
```


## Initial Comparison Code
```{r}
##As an aside compare, whole lake to lake center
center <- read_feather(paste0('out/TS_Preds/NLA2012_cntr_',iteration,'.feather')) %>% mutate(sample = 'NLA') %>%
  bind_rows(read_feather(paste0('out/TS_Preds/Over10_',iteration,'.feather')) %>% mutate(sample = 'Over10')) %>%
  bind_rows(read_feather(paste0('out/TS_Preds/EcoReg2000_cntr_',iteration,'.feather')) %>%
              mutate(sample = 'Random')) %>%
  filter(month %in% c(8:9))

centerComps <- center[,c('COMID', 'date', 'value')] %>% rename(center = value) %>% left_join(predicted[,c('COMID', 'date', 'value')] %>% rename(wholeLake = value)) %>% na.omit()

ggplot(centerComps, aes(x = center, y = wholeLake)) + 
  geom_hex() + scale_fill_viridis_c() +theme_bw() + 
  labs(x = 'Center Estimate', y = 'Whole Lake Estimate', title = 'Center Pixel vs. Whole Lake Comparison', subtitle = '(red line is 1:1)') + 
  geom_abline(color = 'red') +
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3)

## Compare individual preds against each other.
sum.means <- predicted %>%
  #select(-geometry) %>%
  filter(month %in% c(8,9)) %>%
  group_by(COMID, year) %>%
  summarise(Topp.mean = mean(value))


LBPreds <- st_read('in/MN_LB_Comparison/lakesmnwq_1975_2008_fill1975-epsg4326/MNWQ_1975_2008_fill1975-epsg4326.shp')

LBPreds <- LBPreds %>% select(grep('MEAN', names(LBPreds), value = T)) %>%
  st_centroid() %>%
  st_join(ecoregs, join = st_nearest_feature)

plan(multiprocess)
centerJoin <- nhdOut %>% st_transform(crs = 4326) %>%
  split(., c(1:10)) %>%
  future_map(~st_join(.,LBPreds), .progress = T)
plan(sequential)

centerJoin <- rbind(centerJoin[[1]], centerJoin[[2]], centerJoin[[3]], centerJoin[[4]], centerJoin[[5]],
                    centerJoin[[6]], centerJoin[[7]], centerJoin[[8]], centerJoin[[9]], centerJoin[[10]])


LBjoin <- centerJoin %>% select(COMID,grep('MEAN', names(LBPreds), value = T)) %>%
  st_set_geometry(NULL) %>%
  gather(X1975MEAN:X2008MEAN, key = 'year', value = 'LB.mean') %>%
  mutate(year = ifelse(year == 'X1975MEAN', 1975,
                       ifelse(year == 'X1985MEAN', 1985,
                              ifelse(year == 'X1990MEAN', 1990,
                                     ifelse(year == 'X1995MEAN', 1995,
                                            ifelse(year == 'X2000MEAN', 2000,
                                                   ifelse(year == 'X2005MEAN', 2005,
                                                          ifelse(year == 'X2008MEAN', 2008,NA))))))),
         COMID = factor(COMID)) %>%
  filter(LB.mean != 0)
  
  
obs.join <- sum.means %>% left_join(LBjoin)
#write_feather(obs.join, 'out/MN_LB_Comparison/FullObsJoin.feather')
obs.join <- read_feather('out/MN_LB_Comparison/FullObsJoin.feather')

ggplot(obs.join, aes(x = LB.mean, y = Topp.mean)) + geom_hex() + 
  scale_fill_viridis_c() +theme_bw() + 
  #labs(x = 'Center Estimate', y = 'Whole Lake Estimate', title = 'Center Pixel vs. Whole Lake Comparison', subtitle = '(red line is 1:1)') + 
  geom_abline(color = 'red') +
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  coord_cartesian(xlim = c(0,10), ylim = c(0,10)) +
  facet_wrap(~year) + 
  labs(x = 'LakeBrowser Prediction', y = 'This Study Prediction') 

ggsave('figures/__MN_LakeBasedComps.png', width = 5.75, height =3.5, units = 'in')

yearlyMetrics <- obs.join %>%
  na.omit() %>%
  group_by(year) %>%
  summarise(bias = Metrics::bias(LB.mean, Topp.mean),
            mae = mae(LB.mean, Topp.mean))
yearlyMetrics %>% kable() %>% kable_styling()

## See what in situ data we've got in AquaSat
check <- srMunged %>% filter(COMID %in% LBjoin$COMID, 
                             month %in% c(8,9),
                             year %in% c(1985,1990,1995,2000,2005,2008)) %>%
  select(COMID, date, year, value) %>%
  group_by(COMID, year) %>%
  summarise(field = mean(value))

fieldCheck <- check %>% inner_join(obs.join)

p1 <- fieldCheck %>% ggplot(., aes(x = LB.mean, y = field))  + geom_hex() + 
  facet_wrap(~year) +
  scale_fill_viridis_c(trans = 'log10') +
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  geom_abline(color = 'red') +
  theme_bw() +
  #coord_cartesian(xlim = c(0,10), ylim = c(0,10)) + 
  labs(tag = 'A)', x = 'LakeBrowser Mean', y = 'Field Mean')  

p2 <- fieldCheck %>% ggplot(., aes(x = Topp.mean, y = field)) + geom_hex() + 
  facet_wrap(~year) +
  scale_fill_viridis_c(trans = 'log10') +
  stat_poly_eq(aes(label =  paste(stat(adj.rr.label))),
               formula = y~x, parse = TRUE, 
               label.y = Inf, vjust = 1.3) +
  geom_abline(color = 'red') +
  theme_bw() +
  #coord_cartesian(xlim = c(0,10), ylim = c(0,10)) +
  labs(tag = 'B)', x = 'This Study Mean', y = 'Field Mean')

g <- grid.arrange(p1,p2, nrow = 2)

ggsave('figures/__LBCompFieldVal.png', plot = g, width = 6, height = 8)

p3 <-fieldCheck %>%
  na.omit() %>%
  group_by(year) %>%
  summarise(Field.Topp =round(bias(field, Topp.mean),3),
            Field.LB= round(bias(field, LB.mean),3))%>%
            #pbiasTopp = Metrics::percent_bias(field, Topp.mean),
            #pbiasLB= percent_bias(field, LB.mean)) %>%
  gather(Field.Topp:Field.LB, key = 'Source', value = 'Bias') %>%
  ggplot(., aes(x = year, y = Bias, color = Source)) + geom_point() + geom_line() +
  theme_bw() + ggtitle('LB Bias Comp (Mean Summer Preds vs AS Field Means)')


mn <- st_as_sf(maps::map("state", plot = FALSE, fill = TRUE)) %>% filter(ID == 'minnesota')

p4 <- output.full %>% st_as_sf(coords = c('long', 'lat'), crs = 4326) %>% 
  st_join(mn) %>% select(-geometry) %>% filter(!is.na(ID)) %>%
  group_by(year) %>% summarise(bias = bias(Actual, Predicted)) %>% ggplot(aes(x = year, y = bias)) + geom_point() + geom_line() + geom_smooth(method = 'lm') +
  theme_bw() + ggtitle('MN Model Validation Bias (Observation Based)') +
  ggpubr::stat_regline_equation(label.x.npc = 'middle', label.y.npc  = 'top')

p5 <- output.full %>% group_by(year) %>% summarise(bias = bias(Actual, Predicted)) %>% ggplot(aes(x = year, y = bias)) + geom_point() + geom_line() + geom_smooth(method = 'lm') +
  theme_bw() + ggtitle('Overall Model Validation Bias (Observation Based)') +
  ggpubr::stat_regline_equation(label.x.npc = 'middle', label.y.npc  = 'top')

g <- grid.arrange(p3, p4, p5)

ggsave('figures/__LBCompBiasComp.png', plot = g, width = 4, height = 5)

```
